Fixed mistakes - float

+++ mcc_avg.py
@@ -25,9 +25,9 @@ file_lines = read_file(sys.argv[1])
 mcc_results = []
 for line in file_lines:
    line_infos = line.split()
-   mcc_int = int(line_infos[1])
+   mcc_int = float(line_infos[1])
    mcc_results.append(mcc_int)
 
-mcc_avg = int(sum(mcc_results)) / int(len(mcc_results))
+mcc_avg = (sum(mcc_results)) / (len(mcc_results))
 
 print (mcc_avg)
